Add higher timeout to jobs
Is this bad practice, yes, yes it is.
I made this quick fix to make sure that all jobs finish.
It was hitting the timeout because the calculation was genuienly taking
too long.
In the future the job should be optimised and hopefully have this
timeout lowered or entirely removed.